提供登录校验功能,跨域及jwt

+++ src/main/java/com/web/pojo/Admin.java
@@ -12,6 +12,6 @@ public class Admin {
     private String username;
     private String password;
     private String phone;
-    private Byte role;
+    private Short role;
 }
